applying patches from tracker (5 Issues working with files bigger than 4Gb - ID: 3312690) - patch 01-IntOverflow-Filter_Majority.patch

+++ saga-gis/src/modules/grid/grid_filter/Filter_Majority.h
@@ -93,7 +93,7 @@ protected:
 
 private:
 
-	int						m_Radius, m_Threshold;
+	long					m_Radius, m_Threshold;
 
 	CSG_Grid				m_Kernel, *m_pInput;
